解决了加噪问题

+++ DPsys006/src/main/java/hnu/mn/utils/AddLaplaceNoise.java
@@ -3,7 +3,7 @@ package hnu.mn.utils;
 import org.apache.commons.math3.distribution.LaplaceDistribution;
 
 public class AddLaplaceNoise {
-	public static double laplaceMechanismCount(long realCountResult, double epsilon) {
+	public static double laplaceMechanismCount(double realCountResult, double epsilon) {
 		LaplaceDistribution ld = new LaplaceDistribution(0, 1 / epsilon);
 		double noise = ld.sample();
 		return realCountResult + noise;
